fix: v3.0.0 linkified headings don't play well with vuepress. Closes #48.
Term indexing has been completely rewritten and glossarify-md now uses
AST-based indexing and index file generation. With the new dedicated
'indexer' unifiedjs processor plug-in we no longer need to do indexing
and linking in the same processor plug-in.

This caused the bug: in v3.0.0 we did linking and indexing in the same
processor plug-in which required us to run 'remark_link_headings' prior
to running our 'linker'. Only this way we had the heading link information
available which we need for creating links from the index file to those
sections.

However, this brought us in a bad situation where we had to apply
'remark_link_headings' prior to the 'linker' and 'remark_ref_links'
after the 'linker'. In this situation markdown headings were alway
converted to reference link syntax '[My Heading][1]` which vuepress or
'vuepress-bar' didn't recognize and kept in the eventual HTML results
(GitHb was fine, though).

The new implementation with the separate processor plug-in is now
much cleaner and allows us run the 'linker' step first, then convert
any generated glossary links via 'remark_ref_links' and after that
linkify headings into '[My Heading](#my-heading)' syntax and then
apply the 'indexer'.