Drop unnecessary tree traversal.
The previous traversal had the only purpose to mark nodes
which shall be ignored in the subsequent flat mapping.
However since the flatMap callback receives all the args
as the visitor callback we can check parenthood infos
in flatMap as well and skip the unnecessary traversal
as well as adding and deleting _ignoreMe marker property.