feat: Improved performance
* Calculate a term's short description only once.

* Read and sort term definitions only once.

After glossaries have been parsed the terms extracted
are stable. There is no need to read (and sort) terms
alphabetically on each node processed as part of link
insertion.

* Disable stringifcation when reading glossaries.

Reading glossaries has only the purpose of building
the dictionary. The actual markdown output will be
written as part of markdown document processing
afterwards which includes glossaries a second time
for linkification inside of glossaries.

* Drop unnecessary tree traversal.

The previous traversal had the only purpose to mark nodes
which shall be ignored in the subsequent flat mapping.
However since the flatMap callback receives all the args
as the visitor callback we can check parenthood infos
in flatMap as well and skip the unnecessary traversal
as well as adding and deleting _ignoreMe marker property.