test: Config to test another test case.
There is already a test case with `includeFiles: ["."]` where an array
has a single element. In this case we expect that non-default value to
be taken and further expect that merging with the default value '["."]'
does NOT _add_ the default value. So this result were wrong either:
`includeFiles: [".", "."]`.

By emptying glossaries we have another test case where the array is empty
and where we expect the default array value to be added.