Migrating to remark@13.0.0 with remark-parse@9.0.0 and new micromark parser (#132)
BREAKING CHANGES: With remark-parse having switched to a
completely new yet CommonMark compatible markdown parser
there's a (minor) risk that output produced by glossarify-md
changes. It mostly changes in a way which makes the output be
more compliant with the CommonMark spec. Based on what we
observed by comparing outputs with our previous baseline there
were only marginal changes. These are the ones we accepted as
part of our new test baseline:

**Indentation**: Remark strips leading spaces on new lines
either because there is no syntax construct which requires
them or to correct indentation, e.g. of list items.

**Headings**: Remark lifts headings with a depth > 6 into
the valid range of 1-6 according to CommonMark Spec v0.29.
So inputs with headings at a depth of 7 or more are being
lifted to depth 6 by remark (and hence glossarify-md).

**Quotation marks**: In the past it was possible to mix
double quotes and single quotes in a single sentence.
With this update single quotes will be converted into
double quotes and properly escaped where necessary.

**Escapes**: There may be a few changes to what is being
escaped by a leading backslash (Markdown character escape
sequence).

**Newlines**: Newlines in a glossary term's paragraph are
propagating to the glossary link definition created for a
term occurrence. Those are created when linking a term
occurrence with the term definition in a glossary. But
newlines could have propagated for other reasons already
in the past, so it is a very minor change and nothing to
worry about.

The remark changelog can be found here:
https://github.com/remarkjs/remark/releases/tag/13.0.0