feat: 'glossaries[i].file' supports glob patterns ('termHint' and 'title' will be ignored).
If 'file' is a glob pattern 'termHint' and 'title' will be ignored.
There are various reasons for this.

First is performance. For each file terminator had to scan the
'glossaries' array for entries with a glob pattern and test
whether the current entry is the one matching the file being
processed in order to apply the correct 'termHint' and 'title'.

Second: if there were two patterns matching the same file it
would not be possible to decide which 'termHint' to apply.
Basically 'termHint' loses its original purpose, anyways, if
there are multiple files using the same term hint.

Third: glob patterns are being introduced to support cross-linking.
Unlike the option's name suggests documents being matched are no longer
glossaries, only but any document. Headings are no longer terms
but headings being processed like terms. So 'termHint' isn't a hint
on a term anymore.