feat: New option 'unified' for extended markdown processing configuration (#140)
This commit opens up a whole lot new possibilities for users of glossarify-md.
They are now able to configure the 'unified' processing pipeline themeselves
using the unified's configuration format. This in particular means they'll be
able to install and load remark plug-ins and process markdown files with
third-party syntax extensions to the extend supported by unified's, remark's
and micromark's ecosystem.