chore(refactor): Infer TermDefinition anchor from `heading` id
[remark-autolink-headings](https://github.com/remarkjs/remark-autolink-headings)
is considered legacy. Currently glossarify-md's *terminator* expects headings to
be linkified prior to termination in order to read a `link` node's url and pass
it as a TermDefinition's *anchor*. But this is not necessary, actually, due to
the fact that the plug-in, internally, does only produce URLs from `heading`
node identifiers that have been created by `remark-slug` which we have in our
pipeline anyways. So *terminator* could directly inspect a `heading` node's id
and use this one for a TermDefinition anchor. Then it would no longer depend on
headings being linkified.

Note that this commit focuses on termination. It may still not be possible to
drop remark-autolink-headings. This is a non-breaking internal refactoring.