fix(perf): Unnecessary "unwrap" iterations.
Previously when linkifying the documents we iterated on every
document on every paragraph (or tableCell) all the terms
_and_ for each term unwrapped all its existing definitions
from an IndexEntry to the actual TermDefinitionNode ...before
beginning to linkify the paragraph.

The total number of unwrap operations could scale
according to this equation:

~~~
documents * paragraphs * terms * definitions
~~~

After this change the same code section scales
according to

~~~
documents * paragraphs * terms
~~~

In a typical "glossary" use case there's just one
glossary and each term is likely to have only 1
definition. There won't be any real improvements for
these usage scenarios.

But glossarify-md may also be used to cross-reference
a complete book or (technical) documentation with many
pages (using *globs*). In technical docs there are
often pages which are structured according to a particular
page template (e.g. Architecture Decision Records).
The templated headings are to be considered terms of
multiple glossaries, so terms with many different
"(re-)definitions".

For these usage scenarios and inputs we expect a
measurable performance improvement once the number of
files reaches a certain threshold.