feat(compatibility): Drop deprecated `url.parse()` NodeJS-API. (#178)
BREAKING CHANGE: Unfortunately, there is no replacement for
the old API that is fully compatible. As a consequence this
change affects the way paths are being URL-encoded. Previously
all paths (relative or absolute) have been passed through
`url.parse(path).format()` and thereby got URL-encoded. With
this change encoded URLs are only created for option
`paths: "absolute"` in combination with a `baseUrl` other
than "/". The encoding may changed. Values for `baseUrl`
(if present) now *must* be "/" or *must* be URLs which
conform to WHATWG URL specification and which can be passed
as a 'base' parameter to Node's URL implemenation
(see https://docs.nodejs.org/api/url.html).

* test: New baseline.