feat: Apply options like 'termHint' and 'sort' on files matched by glob pattern, too.
Each 'glossaries' entry describes a file _set_. Entries with a 'file' path
describe a set with a single item. Entries with a glob pattern describe
a set with potentially many items.

If there are multiple 'glossaries' entries and therefore multiple file sets
and file sets overlap (that is a file becomes part of multiple sets) then
options of the 'glossaries' entry that is latest in the 'glossaries' array
apply ("Override semantics").