fix: headingIdAlgorithm: No fileset unique hashes with glossary uri.
BREAKING CHANGE: Affects users of 6.0.0-alpha and 6.0.0-beta.0
defining a `glossaries[].uri` and using `headingIdAlgorithm`
`md5` or `sha256` since it changes hashsums and a term's URI.

Previously a term's hashsum was created from a term's plaintext
URI consisting of vocabulary-URI + slug. But when there's a
'glossaries' entry with a glob pattern and an 'uri' then once
the same term existed twice or more in different matched glossary
files their plaintext URI and hashsums would collide. The change
fixes this by including the glossary files actual path relative
to 'baseDir'.