Refactor: Hierarchical histograms.
When disambiguating ambiguous terms in parent sections take
terminology use in subsections into account. Before this
change parent sections have been evaluated in isolation
ignoring glossary references in subsections. For example
when using ambiguous terms in an introductory paragraph
of a chapter disambiguation did only look at glossary
popularity in the introductory paragraph of the chapter
ignoring glossary popularity in subsections of the same
chapter. Now it disambiguates parent sections based on
an aggregation of glossary references in the introductory
paragraph and also  all subsections of the chapter.